+++ internal/checks/base_test.go
@@ -62,8 +62,9 @@ func simpleProm(name, uri string, timeout time.Duration, required bool) *promapi
 	return promapi.NewFailoverGroup(
 		name,
 		[]*promapi.Prometheus{
-			promapi.NewPrometheus(name, uri, map[string]string{"X-Debug": "1"}, timeout, 16, 1000, 1000),
+			promapi.NewPrometheus(name, uri, map[string]string{"X-Debug": "1"}, timeout, 16, 1000),
 		},
+		1000,
 		required,
 	)
 }
@@ -118,7 +119,7 @@ func runTests(t *testing.T, testCases []checkTest) {
 
 			prom := tc.prometheus(uri)
 			if prom != nil {
-				prom.StartWorkers()
+				prom.StartWorkers(time.Hour)
 				defer prom.Close()
 			}